backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Compress==1.14
Flask-Limiter==3.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
Pillow==10.1.0
APScheduler==3.10.4
gunicorn==21.2.0
